refactor: Change auth dto path
